add Crashlytics
Crashlytics is a really strong framework to keep check on your App's health.
You can receive the key after creating an account with https://get.fabric.io/
Once you have the key, add it to the `local.properties` as below

    crashlytics.key=<Your Crashlytics Key Here>

It is always a good idea to keep all the sensitive keys in `local.properties`, as this file is never committed to the repo.

Inside the App's `build.gradle`, we will parse this property and make it available to `AndroidManifect.xml`

    manifestPlaceholders = [crashlytics_key: properties.get("crashlytics.key")]

Once we parse the key, we can access it in AndroidManifest as `${crashlytics_key}`

At this instance, lets also give our app a sensible name.
Crashlytics will record all the events generated by the app in production as well as development. So it is very important that we identify both the versions separately.

For doing this, we will add an `applicationIdSuffix` to the debug variant, this will force Fabric to recognize it as a separate app.
We will also generate the `app_name` separately for `debug` and `release` variants to apmlify the difference.
We will use Gradle's DSL for generating *string resources* in the App's `build.gradle`

    buildTypes {

        debug {
            applicationIdSuffix ".dev"
            resValue "string", "app_name", "CC Player DEV"
        }

        release {
            ...
            resValue "string", "app_name", "CC Player"
        }
    }

This will automatically generate `@string/app_name` resource.
We should now remove the `app_name` string recource from `values/strings.xml`, otherwise Android Studio will throw a `Duplicate recources` error

Let's go ahead and initialize Crashlytics inside `App.kt`

    // Initialize Crashlytics
    Fabric.with(this, Crashlytics())

If you now run the app, Fabric will automatically create a new App entry on your dachboard.